Calculate mean expression values from all expression data with replicates

Calculate mean expression values from all expression data with replicates

Usage

get_mean_data(
  exp,
  expression_value_threshold = 5,
  accession_data_to_transform,
  is_data_normalised = FALSE
)

Arguments

exp

Input data frame containing all replicates of gene expression in each genotype at each time point.

expression_value_threshold

Expression value threshold. Remove expressions if maximum is less than the threshold. If NULL keep all data.

accession_data_to_transform

Accession name of data which will be transformed.

is_data_normalised

TRUE if dataset has been normalised prior to registration process.

Value

A data frame contains only mean expression data.