Register or synchronize different expression profiles

register() is a function to register expression profiles a user wishes to compare.

Usage

register(
  input,
  stretches = NA,
  shifts = NA,
  reference,
  query,
  overlapping_percent = 0.5,
  optimise_registration_parameters = TRUE,
  optimisation_config = list(num_iterations = 60)
)

Arguments

input

Input data frame containing all replicates of gene expression in each genotype at each time point.

stretches

Candidate registration stretch factors to apply to query data, only required if optimise_registration_parameters = FALSE.

shifts

Candidate registration shift values to apply to query data, only required if optimise_registration_parameters = FALSE.

reference

Accession name of reference data.

query

Accession name of query data.

overlapping_percent

Number of minimum overlapping time points. Shifts will be only considered if it leaves at least these many overlapping points after applying the registration function.

optimise_registration_parameters

Whether to optimise registration parameters with Simulated Annealing. By default, FALSE.

optimisation_config

List with optional arguments to modify the Simulated Annealing optimisation.

Value

This function returns a list of data frames, containing:

data

a table containing the scaled input data and an additional timepoint_reg column after applying registration parameters to the query data.

model_comparison

a table comparing the optimal registration function for each gene (based on all_shifts_df scores) to model with no registration applied.

Examples

if (FALSE) {
# Load a data frame from the sample data
data_path <- system.file("extdata/brapa_arabidopsis_all_replicates.csv", package = "greatR")
all_data <- utils::read.csv(data_path)

# Running the registration
registration_results <- register(
  input = all_data,
  reference = "Ro18",
  query = "Col0"
)
}